// desk_docviewer/ui/src/mock/type.ts
export interface Effect {
  description?: string;
  name?: string;
  type?: string;
}
// ...
export interface Companion {
  actions?: {
    [k: string]: BaseAction1;
  };
  description?: string;
  effects?: Effect[];
  skills?: string[];
}
export interface BaseAction1 {
  damage?: number;
  damage_type?: string;
  description?: string;
  duration?: string;
  gauge?: number;
  maintained?: boolean;
  major?: number;
  minor?: number;
  nature?: "physique" | "mana" | "duale";
  range?: "contact" | "LDV" | "perso";
  score?: number;
  threshold?: number;
  type?: string;
  zone?: boolean;
}